# More efficient object finalising implementation