--- a/src/libponyrt/gc/objectmap.h
+++ b/src/libponyrt/gc/objectmap.h
@@ -9,7 +9,6 @@ PONY_EXTERN_C_BEGIN
 typedef struct object_t
 {
   void* address;
-  pony_final_fn final;
   size_t rc;
   uint32_t mark;
   bool immutable;
@@ -22,12 +21,7 @@ object_t* ponyint_objectmap_getobject(objectmap_t* map, void* address, size_t* i
 object_t* ponyint_objectmap_getorput(objectmap_t* map, void* address,
   uint32_t mark);
 
-object_t* ponyint_objectmap_register_final(objectmap_t* map, void* address,
-  pony_final_fn final, uint32_t mark);
-
-void ponyint_objectmap_final(objectmap_t* map);
-
-size_t ponyint_objectmap_sweep(objectmap_t* map);
+void ponyint_objectmap_sweep(objectmap_t* map);
 
 PONY_EXTERN_C_END