@jemc 2017-03-06
I think this should be split into two PRs, especially since the changes don't seem to have a strict dependency relationship to eachother (for example, the finaliser change doesn't strictly depend on having better benchmark facilities.

I'm also inclined to say the `ponybench` changes should go through the RFC process to get more feedback about the design, but maybe somebody else will disagree with me about that and we can discuss.