@dipinhora 2017-03-07
@jemc I'm okay with splitting this PR into two PRs.

I only made the `ponybench` changes because the process got `OOMKilled` otherwise:

```
vagrant@buffy-leader-1:~/hp2$ ./finalisers
primitive init
Killed
```

I modeled the `ponybench` changes based on how `ponytest` works (except that I used `options` which `ponytest` explicitly avoids).

If the `ponybench` changes do go through an `RFC`, it'd be a good idea to discuss whether having `ponybench` run all iterations in a single behavior is desired or not (or if it should be dynamically choosable as a runtime option). The current design of having all iterations run in a single behavior is why the `finalisers` benchmark was getting `OOMKilled`, however, I do understand that running iterations in multiple behavior calls will result in the added overhead of GC and contention from other actors so it's not a simple decision.

Regardless, I'll split this PR and/or open the RFC once other folks have had a chance to weigh in (no point in opening another PR if the decision is to go through an RFC first).