@jemc 2017-03-07
> it'd be a good idea to discuss whether having ponybench run all iterations in a single behavior is desired or not (or if it should be dynamically choosable as a runtime option).

Yeah, I agree that sounds to me like the good topic of an RFC discussion.