@dipinhora 2017-03-09
@Praetonus @jemc Are there any other comments, questions or changes for this PR before it's merged?

I'll try and get the `ponybench` RFC started sometime in the next few days.