@Praetonus 2017-03-09
I'm not sure about the benchmark added to `minimal-cases/finalisers`. If I recall correctly, the minimal cases directory was added to contain regression tests that couldn't be added as compiler unit tests because we couldn't yet run a Pony program inside of a compiler unit test. Now that we can do that, the minimal cases should progressively move to the compiler test suite (and I've already started that with finalisers in #1629), so I don't think we should add more things in the files. Is there another place to put the benchmark?