@dipinhora 2017-03-09
@Praetonus I understand your concern and agree that there should be another place for the benchmarks. However, the only other Pony based benchmarks in the repo are part of `stdlib` for the `persistent` maps in `packages/collections/persistent/benchmarks/`.

Any suggestions for where to put this benchmark since it's not testing a specific `stdlib` package?