@Praetonus 2017-03-09
Maybe it could be implemented in C and added to `benchmark/libponyrt`?