@dipinhora 2017-03-10
@Praetonus I've updated the PR to remove the changes to `minimal-cases/finalisers` so that this PR can be merged without a benchmark.

I'll work on creating a benchmark in C that exercises the same functionality to add into `benchmark/libponyrt` and submit that as another PR when it is ready.