@sylvanc 2017-03-10
Sorry to be a pain, @dipinhora , but I think another set of changes is needed. In `heap.c`, you are passing a `has_finaliser` flag to several methods. This results in an extra test and branch instruction on every allocation.

However, whether a finaliser is present or not is known statically when the memory allocation occurs, so we don't need to pay this cost.

If we separate out functions that set the `finalise` bit from those that don't in `heap.c`, we can statically call the right one, and eliminate the branch during allocation.