@dipinhora 2017-03-10
@sylvanc No problem at all and good catch on the extra/unnecessary branch. I didn't even notice it. 8*/

I've eliminated branch by converting the `has_finaliser` to a `uint32_t` which should turn `true` to `1` and `false` to `0` since we're using C99 `stdbool.h`.

Please let me know if you find anything else that needs fixing.