@sylvanc 2017-03-11
Ah I see what you did there :) But what I meant was, just like how in `pony.h` we have separate entry points for allocators with finalisers and without, we should maintain that distinction in `heap.c`. That avoids paying _any_ cost, even bit shifts, for allocations of objects that have no finalisers. As such, there is no performance drop (not even a tiny one) on alloc compared to the old code.