@dipinhora 2017-03-11
@sylvanc changes made to add functions `ponyint_heap_alloc_final`, `ponyint_heap_alloc_small_final` and `ponyint_heap_alloc_large_final` that are a copy of the logic from the original functions. This introduces the maintenance overhead of making sure all core logic changes to `ponyint_heap_alloc_small` are also made in `ponyint_heap_alloc_small_final` and vice versa. Please let me know if you had something other than copying the logic in mind.

Also, one minor note, there is still a tiny overhead with the new implementation of alloc (without a finaliser) as compared to the old code because the new code needs to set `chunk->finalisers` that the old code didn't. I can't think of any way to avoid that without adding in a `memset` to `0` the new `chunk_t` first which would also add overhead as compared to the old code.